+++ Inheritance_HW3/FamilyApartment.h
@@ -4,7 +4,7 @@
 class FamilyApartment : public Flat		//CsaladiApartman
 {
 public:
-	FamilyApartment(const int& area, const int& rooms, const int& costpersqrmtr, const int& people=0, const int& children=0);
+	FamilyApartment(const double& area, const int& rooms, const int& costpersqrmtr, const int& people=0, const int& children=0);
 
 	bool move_in(const int& noOfPeople) override;
