+++ Inheritance_HW3/Garage.h
@@ -6,7 +6,7 @@
 class Garage : public Rentable, public RealEstate
 {
 public:
-	Garage(const int& area, const int& costprsqm, const bool& heated, const bool& car=0, const int& reservedfor=0 );
+	Garage(const double& area, const int& costprsqm, const bool& heated, const bool& car=0, const int& reservedfor=0 );
 
 	int get_cost() const override;
 	int total_cost(const int& numberOfMonths) override;
@@ -18,7 +18,7 @@ public:
 	friend std::ostream& operator<<(std::ostream& outputStream, const Garage& garage);
 
 private:
-	int mArea;
+	double mArea;
 	int mCostPerSqrmtr;
 	bool mHeated;
 	bool mParkedIn;
